Fat binary!
Added aggregate target to make creating a fat binary easy.

Also fixed a few classes to use NSInteger instead of int and to fix some int promotion problems with sizeof issues.

+++ HamsterTools/NSArray+ArrayUtils.m
@@ -12,15 +12,14 @@
 
 -(id)objectAtIndexedSubscript:(NSInteger)idx {
     
-    NSUInteger index = idx;
-    
+    NSUInteger index = (NSUInteger)idx;
+      
     if (idx < 0) {
         // Negative index.
-        index = self.count + idx;
+        index = self.count + (NSUInteger)idx;
     }
     return [self objectAtIndex:index];
 }
-
 
 
 @end
